+++ src/webgis4u/util.js
@@ -3,6 +3,8 @@
  */
 
 export * from './util/debounce';
+export * from './util/dom';
 export * from './util/formatNumber';
 export * from './util/formatText';
 export * from './util/pan';
+export * from './util/string';